SQUASH: AK3: Use proper EAS checker algorithm
I don't want to mess with blindly copying powerhint.json file to ROM that not have
that file or HMP based ROM, so we need to check if that ROM have that json file or not
as parameter to EAS Based ROM for now.

This commit remove all ROM name check that's too tricky and may broke if user update build date.

squashed commit:
- https://github.com/Nicklas373/AnyKernel3/commit/062c0c74f704f864fffe4d593073a5f6002a5fde
- https://github.com/Nicklas373/AnyKernel3/commit/0c8bde437ccb6a049a49af892003888446b5eef0

Signed-off-by: Dicky Herlambang <herlambangdicky5@gmail.com>